improve documentation and move frame utility functions  to separate file
