+++ src/common/mod.rs
@@ -1,4 +1,7 @@
 //! This module contains error handling utilities for the application.
 //!
-//! It provides the custom error type `MyError` and various error message constants.
+//! It consists of the following sub-modules:
+//! - `errors`: Defines an `ApplicationError` enum and related functionality for handling application errors.
+//! - `utils`: Contains utility functions
 pub mod errors;
+pub mod utils;
